+++ src/com/dannycrafts/myTitles/database/Row.java
@@ -17,11 +17,8 @@ public class Row
 	
 	public void readCell( int index, Cell cell )
 	{
-		byte[] cellBuffer = new byte[header.getCellLength( index )];
-		System.arraycopy( buffer, header.getCellOffset( index ), cellBuffer, 0, header.getCellLength( index ) );
-		
-		//System.out.println( buffer.length + " - " + cellBuffer[0] );
-
+		byte[] cellBuffer = new byte[header.getCellLength( (short)index )];
+		System.arraycopy( buffer, header.getCellOffset( (short)index ), cellBuffer, 0, header.getCellLength( (short)index ) );
 		cell.write( cellBuffer );
 	}
